requirements.txt
torch
datasets
transformers
evaluate
numpy
Pillow